(BSR)[API] refactor: move get_offers_for_my_institution in educational repository
